% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_DEtest.R
\name{DEtest}
\alias{DEtest}
\title{Identification of differentially regulated metabolomics variates}
\usage{
DEtest(
  outcome,
  features,
  confounders = NULL,
  MWSL = NULL,
  alpha = NULL,
  vennPlot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{\code{outcome}}{a vector or a data frame of \code{n} samples. Possible outcome types are continuous (both symmetric and skewed), discrete binary (0/1), discrete countable, or survival time-to-event}

\item{\code{features}}{a data frame of \code{n} samples (rows) and \code{M} features (columns)}

\item{\code{confounders}}{an optional data frame of \code{n} samples (rows) and \code{P} fixed effects confounders (columns), default=\code{NULL}}

\item{\code{MWSL}}{metabolome-wide significance level (MWSL) estimated via the permutation-based approach (see \code{MWSL::FWERperm}) or via the via the closed-form-expression approach (see \code{MWSL::Meff})}

\item{\code{alpha}}{an optional probability value, default=0.05}

\item{\code{vennPlot}}{an optional logic value to be set to \code{TRUE} for the visualisation of the Venn-plot for the total count of differrentally regulated metabolomics variates when employing a Bonferroni or Benjamini-Hochberg (FDR) or the MWSL (FWER) correction, respectively}

\item{\code{verbose}}{an optional logic value to suppress some output status messages, default=\code{TRUE}}
}
\value{
\code{res.DE_count} total count of differrentally regulated metabolomics variates when employing a Bonferroni or Benjamini-Hochberg (FDR) or the MWSL (FWER) correction, respectively

\code{res.DE_names} names of the of differrentally regulated metabolomics variates when employing a Bonferroni or Benjamini-Hochberg (FDR) or the MWSL (FWER) correction, respectively

\code{Venn_plot} Venn-plot visualisation tool for the total count of differrentally regulated metabolomics variates when employing a Bonferroni or Benjamini-Hochberg (FDR) or the MWSL (FWER) correction, respectively
}
\description{
Identification of differentially regulated metabolomics variates directly linked to a specified outcome (e.g. disease risk). This includes the case of continuous outcomes both symmetric and skewed, as well as the case of a binary (0/1) outcome, and a countable outcome taking discrete values 0,1,2,3,$\ldots$, as well as a survival time-to-event outcome. Once the appropriate analysis has been performed the metabolites of interest can be identified comparing the respective p-value to the the adjusted thresold (MWSL or Meff). When the raw p-value corresponding to a certain feature is smaller than the adjusted thresold we identify that metabolomic variate as significant. For the other metabolites we conclude that there is no association between the changes in the outcome variable and the shifts in these features.
}
