% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_FWERperm.R
\name{FWERperm}
\alias{FWERperm}
\title{Estimation of permutation-based metabolome-wide significance level (MWSL) and corresponding effective number of tests (ENT)}
\usage{
FWERperm(
  outcome,
  features,
  confounders = NULL,
  n.permutation = NULL,
  method = NULL,
  alpha = NULL,
  verbose = TRUE
)
}
\arguments{
\item{\code{outcome}}{a vector or a data frame of \code{n} samples. Possible outcome types are continuous (both symmetric and skewed), discrete binary (0/1), discrete countable, or survival time-to-event}

\item{\code{features}}{a data frame of \code{n} samples (rows) and \code{M} features (columns)}

\item{\code{confounders}}{an optional data frame of \code{n} samples (rows) and \code{P} fixed effects confounders (columns), default=\code{NULL}}

\item{\code{n.permutation}}{an optional integer number, default=10000}

\item{\code{method}}{an optional string with possible values \code{identity} (no transformation), \code{mN} (multivariate Normal), or \code{mlogN} (multivariate log-Normal), default=\code{mN}}

\item{\code{alpha}}{an optional probability value, default=0.05}

\item{\code{verbose}}{an optional logic value to suppress some output status messages, default=TRUE}
}
\value{
\code{matPvals} the matrix of p-values

\code{q} the vector of minimum p-values

\code{res} the vector of result estimates
\code{MWSL} = metabolome-wide significance level (MWSL)
\code{MWSL_CI.up} = upper value confidence interval MWSL
\code{MWSL_CI.low} = lower value confidence interval MWSL
\code{ENT} = effective number of tests (ENT)
\code{ENT_CI.up} = upper value confidence interval ENT
\code{ENT_CI.low} = lower value confidence interval ENT
\code{R.percent} = ENT/M

\code{t1err.percent} the type I error rate
}
\description{
Parametric approximation methods based on the (log-)multivariate Normal distribution are employed to estimate a stable permutation-based significance level to be used for multiple tests adjustement in the case of correlated metabolomic variates
}
