% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Meff.R
\name{Meff}
\alias{Meff}
\title{Estimation of closed-form-expression metabolome-wide significance level (MWSL) and corresponding effective number of tests (ENT)}
\usage{
Meff(
  features,
  n.permutation = NULL,
  method = "ecorr",
  big.mat = FALSE,
  alpha = NULL,
  verbose = TRUE
)
}
\arguments{
\item{\code{features}}{a data frame of \code{n} samples (rows) and \code{M} features (columns)}

\item{\code{method}}{a string with possible values \code{ecorr} (empirical correlation), or \code{scorr} (shrinkage correlation), default=\code{ecorr}}

\item{\code{alpha}}{an optional probability value, default=0.05}

\item{\code{big.mat}}{an optional logic value to be set to \code{TRUE} when dealing with a very large set of feature (M>>5000), default=\code{FALSE}}
}
\value{
\code{Meff_Nyholt} closed-form expression of the effective number of tests based on Nyholt (2004)

\code{Meff_Liji} closed-form expression of the effective number of tests based on Li \& Ji (2005)

\code{Meff_Gao} closed-form expression of the effective number of tests based on Gao (2008)

\code{Meff_Galwey} closed-form expression of the effective number of tests based on Galwey (2009)

\code{Meff_Bonferroni} closed-form expression of the effective number of tests based on Bonferroni (1963)

\code{Meff_Sidak} closed-form expression of the effective number of tests based on Sidak (1967)

\code{Meff_MWSL} closed-form expression of the effective number of tests based on Peluso (2020)

\code{res.Meff_MWSL} the vector of result estimates
\code{MWSL} = metabolome-wide significance level (MWSL)
\code{MWSL_CI.up} = upper value confidence interval MWSL
\code{MWSL_CI.low} = lower value confidence interval MWSL
\code{ENT} = effective number of tests (ENT)
\code{ENT_CI.up} = upper value confidence interval ENT
\code{ENT_CI.low} = lower value confidence interval ENT
\code{R.percent} = \code{ENT}/\code{M}
}
\description{
Approximation methods based on the spectral decomposition of the correlation matrix of the metabolomics variates are employed to estimate a closed-form-expression significance level to be used for multiple tests adjustement in the case of correlated metabolomic variates
}
